Fixing locations of tests
